+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   Menu tests added to e2e tests (#224)
+-   New users awaiting approval page and feature (#220)
 
 ### Changed